pyproject.toml
[project]
name = "telebot"
version = "0.1.0"
description = ""
authors = [
    {name = "Đặng Xuân Minh Hiếu",email = "zipdang2004@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-telegram-bot[job-queue] (>=22.6,<23.0)",
    "priceapi @ file:///Users/hieu/uriel/onboard/02_backend/priceapi",
]

[tool.poetry]
packages = [{include = "telebot", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
